Use crossplatform way of comparing r_string (#200)
* Removed disable_serialization wrapper (no use in Intercept)

* Use one crossplatform comparison instead of

multiple #ifdefed ones

* fix all the warnings

* Refactoring

* Refactoring

* Fixed missing break in signing switch case

* fix min/max macro interfering

Fixed build errors with slighly incorrect env setup

* Fixed Eventhandlers not working

Hasher was missing EHType
client_eventhandler func was missing EHType
EHType should be uint8_t

* Added compact_array::front()

* Fix

* fix

* Wrong end iterator offset

* Fux

* Fix signing fails on Windows 7

* Added registered_sqf_function::has_function

* Fix iterator operator return val

* Fix invalid game_data_number/bool new/delete operator definition

* r_string > operator

* On Linux every plugin is core signed for now

* Don't potentially crash if there is no logfile

* optimize CBA XEHs

* Revert "Removed disable_serialization wrapper (no use in Intercept)"

This reverts commit 55e48f67173c51b59c82fb1af7eef619f89eddb6.

* Fix typo

+++ src/host/invoker/eventhandlers.cpp
@@ -72,7 +72,7 @@ namespace intercept {
 
     game_value eventhandlers::client_eventhandler(game_value_parameter left_arg, game_value_parameter right_arg) {
         r_string moduleName = left_arg[0];
-        const int ehType = left_arg[1];
+        const uint8_t ehType = static_cast<int>(left_arg[1]);
         const float uidf = left_arg[2];
         const int32_t uid = static_cast<int32_t>(uidf);
         //uint32_t handle = static_cast<uint32_t>(static_cast<float>(left_arg[3]));
